+++ src/tools.js
@@ -2,22 +2,27 @@ const Apify = require('apify');
 const vm = require('vm');
 const _ = require('underscore');
 const queryString = require('query-string'); // TODO: Use Node default
-const { REQUIRED_PROXY_GROUP } = require('./consts');
+const { REQUIRED_PROXY_GROUP, POSITION_FIELD_NAME } = require('./consts');
 const {
     DEFAULT_GOOGLE_SEARCH_DOMAIN_COUNTRY_CODE,
     COUNTRY_CODE_TO_GOOGLE_SEARCH_DOMAIN,
     GOOGLE_SEARCH_URL_REGEX,
     GOOGLE_DEFAULT_RESULTS_PER_PAGE,
+    RESULT_TYPE,
 } = require('./consts');
 
-exports.createSerpRequest = (url, page) => {
-    if (url.startsWith('https://')) url = url.replace('https://', 'http://');
+const { utils: { log } } = Apify;
+
+exports.createSerpRequest = (url, userData) => {
+    const newUrl = new URL(url);
+    newUrl.protocol = 'http:';
+    newUrl.hostname = newUrl.hostname.startsWith('www.')
+        ? newUrl.hostname
+        : 'www.'.concat(newUrl.hostname);
 
     return {
-        url,
-        userData: {
-            page,
-        },
+        url: newUrl.toString(),
+        userData,
     };
 };
 
@@ -32,11 +37,17 @@ exports.getInitialRequests = ({
 }) => {
     return queries
         .split('\n')
-        .map(item => item.trim())
-        .filter(item => !!item)
+        .map((item) => item.trim())
+        .filter((item) => !!item)
         .map((queryOrUrl) => {
             // If it's search URL ...
-            if (GOOGLE_SEARCH_URL_REGEX.test(queryOrUrl)) return exports.createSerpRequest(queryOrUrl, 0);
+            const userData = {
+                page: 0,
+                organicResultsCount: 0,
+                paidResultsCount: 0,
+            };
+
+            if (GOOGLE_SEARCH_URL_REGEX.test(queryOrUrl)) return exports.createSerpRequest(queryOrUrl, userData);
 
             // Otherwise consider it as query term ...
             const domain = COUNTRY_CODE_TO_GOOGLE_SEARCH_DOMAIN[(countryCode || '').toUpperCase()]
@@ -51,7 +62,7 @@ exports.getInitialRequests = ({
             if (mobileResults) qs.xmobile = 1;
             if (includeUnfilteredResults) qs.filter = 0;
 
-            return exports.createSerpRequest(`http://www.${domain}/search?${queryString.stringify(qs)}`, 0);
+            return exports.createSerpRequest(`http://www.${domain}/search?${queryString.stringify(qs)}`, userData);
         });
 };
 
@@ -83,7 +94,7 @@ exports.getInfoStringFromResults = (results) => {
 };
 
 exports.logAsciiArt = () => {
-    console.log(`
+    log.info(`
  _______  _______  _______  _______  _______ _________ _        _______
 (  ____ \\(  ____ \\(  ____ )(  ___  )(  ____ )\\__   __/( (    /|(  ____ \\
 | (    \\/| (    \\/| (    )|| (   ) || (    )|   ) (   |  \\  ( || (    \\/
@@ -121,7 +132,7 @@ exports.createDebugInfo = (request, response) => {
 exports.ensureAccessToSerpProxy = async () => {
     const userInfo = await Apify.newClient().user().get();
     // Has access to group and nonzero limit.
-    const hasGroupAllowed = userInfo.proxy.groups.filter(group => group.name === REQUIRED_PROXY_GROUP).length > 0;
+    const hasGroupAllowed = userInfo.proxy.groups.filter((group) => group.name === REQUIRED_PROXY_GROUP).length > 0;
     const maxSerps = userInfo.limits
         ? userInfo.limits.monthlyGoogleSerpRequests
         : userInfo.plan.maxMonthlyProxySerps;
@@ -140,4 +151,55 @@ exports.ensureAccessToSerpProxy = async () => {
             + ' Please contact support@apify.com to increase the limit.');
         process.exit(1);
     }
+};
+
+/**
+ *
+ * @param {Apify.Dataset} dataset
+ * @param {Any} results
+ * @param {Boolean} csvFriendlyOutput
+ * @param {{ organicResultsCount: Number, paidResultsCount: Number }} resultsCount
+ */
+exports.saveResults = async (dataset, results, csvFriendlyOutput, resultsCount) => {
+    const datasetResults = csvFriendlyOutput ? buildCsvFriendlyResults(results, resultsCount) : results;
+
+    await dataset.pushData(datasetResults);
+};
+
+const buildCsvFriendlyResults = (results, { organicResultsCount, paidResultsCount }) => {
+    const { organicResults, paidResults, searchQuery } = results;
+
+    const organicAndPaidResults = [
+        ...getTypedResults(paidResults, searchQuery, RESULT_TYPE.PAID, paidResultsCount, POSITION_FIELD_NAME.PAID),
+        ...getTypedResults(organicResults, searchQuery, RESULT_TYPE.ORGANIC, organicResultsCount, POSITION_FIELD_NAME.ORGANIC),
+    ];
+
+    return organicAndPaidResults;
+};
+
+const getTypedResults = (results, searchQuery, type, resultsCount, positionFieldName) => {
+    // We want different position field names for organic and paid results (names are defined in constants).
+    const position = {};
+    position[positionFieldName] = resultsCount + 1;
+
+    const typedResults = results.map((result) => {
+        const typedResult = {
+            searchQuery,
+            type,
+            ...position,
+            ...result,
+        };
+
+        // Exclude siteLinks as it is an array of objects.
+        delete typedResult.siteLinks;
+
+        // Stringify an array of keywords.
+        typedResult.emphasizedKeywords = typedResult.emphasizedKeywords.join(' | ');
+
+        position[positionFieldName]++;
+
+        return typedResult;
+    });
+
+    return typedResults;
 };